Adding move dialog
Added the move dialog and implemented it. Also moved values passed to
views into a dto directory and named them specifically. Every view
expecting data needs an own dto class where everything is passed that is
required by the view.

Also added a ValueText object to be used in ComboBoxes where the value
and text isn't the same.