Fixed small issues in history page and main page
Fixed issues with the level history on loading the main page
after searching and enering another page before returning to
the main page.

Adjusted the reset button to display the previous level on reset and not always the base level on the main page.

Added failure message to the history page if the creation
of a CryptoForest tree fails.